requirements.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
tzdata==2022.7
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
duckdb>=0.9.2
pyarrow>=14.0.1
apache-airflow>=2.7.1
apache-airflow-providers-amazon>=8.0.0
openpyxl==3.1.2
boto3==1.28.44
numpy==1.24.3
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
scipy==1.11.4
